feature(User & Task): create User & Task relations

+++ src/task/task.repository.ts
@@ -1,12 +1,15 @@
 import { AppDataSource } from 'database';
+import { FindOptionsWhere } from 'typeorm';
+// task
 import { TaskEntity } from './task.entity';
-import { buildQueryOptions } from 'common/utils';
-import { CreateTaskDto, UpdateTaskDto } from './task.types';
+import { CreateTaskDto, TaskFindOptions } from './task.types';
+// common
 import {
   Nullable,
   PaginatedResponse,
   QueryParamsDtoSchema,
 } from 'common/types';
+import { buildQueryOptions } from 'common/utils';
 
 class TaskRepositoryClass {
   constructor(
@@ -15,40 +18,52 @@ class TaskRepositoryClass {
 
   async findAll(
     queryParams: QueryParamsDtoSchema,
+    options?: TaskFindOptions,
   ): Promise<PaginatedResponse<TaskEntity>> {
     const { skip, take, where, order } = buildQueryOptions<TaskEntity>(
       queryParams,
       ['title', 'description'],
     );
 
+    if (options?.authorId) {
+      where.push({ authorId: options.authorId });
+    }
+
     const [data, total] = await this.repo.findAndCount({
       where,
       order,
       skip,
       take,
+      relations: options?.relations,
     });
 
     return { data, total };
   }
 
-  create(data: CreateTaskDto): Promise<TaskEntity> {
-    const created = this.repo.create(data);
+  create(data: CreateTaskDto, authorId: number): Promise<TaskEntity> {
+    const created = this.repo.create({ ...data, authorId });
     return this.repo.save(created);
   }
 
-  findOne(id: number): Promise<Nullable<TaskEntity>> {
-    return this.repo.findOneBy({ id });
+  findOne(
+    id: number,
+    options?: TaskFindOptions,
+  ): Promise<Nullable<TaskEntity>> {
+    const where: FindOptionsWhere<TaskEntity> = { id };
+
+    if (options?.authorId) {
+      where.authorId = options.authorId;
+    }
+
+    return this.repo.findOne({ where, relations: options?.relations });
   }
 
-  async update(id: number, data: UpdateTaskDto): Promise<TaskEntity | null> {
-    const updatedTask = await this.repo.preload({ id, ...data });
-    if (!updatedTask) return null;
-
-    return this.repo.save(updatedTask);
+  async update(task: TaskEntity): Promise<TaskEntity | null> {
+    return this.repo.save(task);
   }
 
-  delete(id: number) {
-    return this.repo.delete(id);
+  delete(id: number, authorId: number) {
+    return this.repo.delete({ id, authorId });
   }
 }
